# Conflicts:
#	backend/common/log.py

+++ backend/app/admin/crud/crud_user.py
@@ -25,7 +25,8 @@ from backend.app.admin.schema.user import (
     UpdateUserParam,
 )
 from backend.app.admin.utils.password_security import get_hash_password
-from backend.utils.dynamic_import import import_module_cached
+from backend.common.exception import errors
+from backend.plugin.core import check_plugin_installed
 from backend.utils.serializers import select_join_serialize
 from backend.utils.timezone import timezone
 
@@ -134,13 +135,14 @@ class CRUDUser(CRUDPlus[User]):
         db.add(new_user)
         await db.flush()
 
-        role_stmt = select(Role).where(Role.id.in_(obj.roles))
-        result = await db.execute(role_stmt)
-        roles = result.scalars().all()
+        if obj.roles:
+            role_stmt = select(Role).where(Role.id.in_(obj.roles))
+            result = await db.execute(role_stmt)
+            roles = result.scalars().all()
 
-        user_role_data = [AddUserRoleParam(user_id=new_user.id, role_id=role.id).model_dump() for role in roles]
-        user_role_stmt = insert(user_role)
-        await db.execute(user_role_stmt, user_role_data)
+            user_role_data = [AddUserRoleParam(user_id=new_user.id, role_id=role.id).model_dump() for role in roles]
+            user_role_stmt = insert(user_role)
+            await db.execute(user_role_stmt, user_role_data)
 
     async def add_by_oauth2(self, db: AsyncSession, obj: AddOAuth2UserParam) -> None:
         """
@@ -177,16 +179,17 @@ class CRUDUser(CRUDPlus[User]):
 
         count = await self.update_model(db, user_id, obj)
 
-        role_stmt = select(Role).where(Role.id.in_(role_ids))
-        result = await db.execute(role_stmt)
-        roles = result.scalars().all()
-
         user_role_stmt = delete(user_role).where(user_role.c.user_id == user_id)
         await db.execute(user_role_stmt)
 
-        user_role_data = [AddUserRoleParam(user_id=user_id, role_id=role.id).model_dump() for role in roles]
-        user_role_stmt = insert(user_role)
-        await db.execute(user_role_stmt, user_role_data)
+        if role_ids:
+            role_stmt = select(Role).where(Role.id.in_(role_ids))
+            result = await db.execute(role_stmt)
+            roles = result.scalars().all()
+
+            user_role_data = [AddUserRoleParam(user_id=user_id, role_id=role.id).model_dump() for role in roles]
+            user_role_stmt = insert(user_role)
+            await db.execute(user_role_stmt, user_role_data)
 
         return count
 
@@ -308,16 +311,16 @@ class CRUDUser(CRUDPlus[User]):
         :param user_id: 用户 ID
         :return:
         """
+        if check_plugin_installed('oauth2'):
+            try:
+                from backend.plugin.oauth2.crud.crud_user_social import user_social_dao
+
+                await user_social_dao.delete_by_user_id(db, user_id)
+            except ImportError:
+                raise errors.ServerError(msg='OAuth2 插件用法导入失败，请联系系统管理员')
+
         user_role_stmt = delete(user_role).where(user_role.c.user_id == user_id)
         await db.execute(user_role_stmt)
-
-        try:
-            user_social = import_module_cached('backend.plugin.oauth2.crud.crud_user_social')
-            user_social_dao = user_social.user_social_dao
-        except (ImportError, AttributeError):
-            pass
-        else:
-            await user_social_dao.delete_by_user_id(db, user_id)
 
         return await self.delete_model(db, user_id)
